requirements.txt
pytest==7.1.1
selenium==3.14.0
allure-pytest==2.9.45